+++ package.json
@@ -1,29 +1,29 @@
 {
   "name": "mac-aegis",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.10.4",
-    "@emotion/styled": "^11.10.4",
-    "@fontsource/roboto": "^4.5.8",
-    "@mui/icons-material": "^5.10.6",
-    "@mui/material": "^5.10.6",
-    "@mui/x-date-pickers": "^5.0.12",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "dayjs": "^1.11.7",
-    "firebase": "^9.10.0",
-    "million": "^2.5.4-beta.2",
-    "moment": "^2.29.4",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-router-dom": "^6.4.3",
+    "@emotion/react": "11.10.4",
+    "@emotion/styled": "11.10.4",
+    "@fontsource/roboto": "4.5.8",
+    "@mui/icons-material": "5.10.6",
+    "@mui/material": "5.10.6",
+    "@mui/x-date-pickers": "5.0.12",
+    "@testing-library/jest-dom": "5.16.5",
+    "@testing-library/react": "13.4.0",
+    "@testing-library/user-event": "13.5.0",
+    "dayjs": "1.11.7",
+    "firebase": "9.10.0",
+    "million": "2.5.4-beta.2",
+    "moment": "2.29.4",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-router-dom": "6.4.3",
     "react-scripts": "5.0.1",
-    "sass": "^1.54.9",
-    "sweetalert2": "^11.4.33",
-    "web-vitals": "^2.1.4",
-    "xlsx": "^0.18.5"
+    "sass": "1.54.9",
+    "sweetalert2": "11.7.20",
+    "web-vitals": "2.1.4",
+    "xlsx": "0.18.5"
   },
   "scripts": {
     "start": "craco start",
@@ -51,7 +51,7 @@
     ]
   },
   "devDependencies": {
-    "@craco/craco": "^7.1.0",
-    "standard": "^17.1.0"
+    "@craco/craco": "7.1.0",
+    "standard": "17.1.0"
   }
 }
