+++ src/pages/CaseList/CaseList.js
@@ -9,7 +9,7 @@ import Case from '../../components/Case/Case'
 // Firebase
 import db from '../../utils/firebaseConfig'
 // Custom hook
-import { useGetCases } from '../../customHooks/documentHook'
+import { useGetCases } from '../../customHooks/indexHooks'
 // Context
 import { AuthContext } from '../../context/authContext'
