+++ src/pages/ExaDetail/ExaDetail.js
@@ -3,7 +3,7 @@ import { Navigate, useParams, Link } from 'react-router-dom'
 // Librerías
 import CircularProgress from '@mui/material/CircularProgress'
 // Custom Hooks
-import { useGetCases } from '../../customHooks/documentHook'
+import { useGetCases } from '../../customHooks/indexHooks'
 // Context
 import { AuthContext } from '../../context/authContext'
 // Firebase
