+++ src/pages/Search/Search.js
@@ -13,8 +13,7 @@ import Filter from '../../components/Filter/Filter'
 // Utils
 import handlePaste from '../../utils/handlePaste'
 // Custom hook
-import { useGetCells } from '../../customHooks/dataHook'
-import { useGetCases } from '../../customHooks/documentHook'
+import { useGetCells, useGetCases } from '../../customHooks/indexHooks'
 // Context
 import { AuthContext } from '../../context/authContext'
 // XLSX
